SintPowerCase.jl

This package is meant for reading power system data from different sources into data frames. The format is ispired by the MATPOWER case format. The package will also provide some power system matrices based on the data.

Specifying input data

Input data can be specified in a .toml file. Matrices specifying data can be put in csv files. The path to the csv files are given in the [files] section as follows: [files] bus = "bus.csv" branch = "branch.csv" The supported matrices are the fileds of the Case struct.

Changes compared to the MATPOWER case format

Additional load and generator matrices

In addition to defining loads in the bus matrix. It is also possible to define a load matrix in a similar way as the gen matrix used in the MATPOWER case format.

|ID|bus|customer_type| |–|–-|––––––-| |1 |1 |residential | |2 |1 |public |

|ID|bus|OS|P | |–|–-|–|- | |1 |3 |1 |50| |2 |3 |1 |50| |1 |3 |2 |25| |2 |3 |2 |75|

Storage units

The format also supports adding storage devices to the case. They are specified in the gen matrix by adding the additional columns, Emax the energy capacity and E the energy stored.

External grids or reserves

External grids or reserve connections are indicated by an extra column in the gen matrix named external that can take the value 1 or 0.

Method documentation

SintPowerCase.contingency_matrix

SintPowerCase.create_random_states!

SintPowerCase.get_admittance_matrix

SintPowerCase.get_admittance_matrix

SintPowerCase.get_all_buses_aggregated_power

SintPowerCase.get_bus_row

SintPowerCase.get_dc_admittance_matrix

SintPowerCase.get_gen_bus_power

SintPowerCase.get_gen_buses_power

SintPowerCase.get_gen_buses_power

SintPowerCase.get_gen_indices

SintPowerCase.get_id_idx

SintPowerCase.get_incidence_matrix

SintPowerCase.get_load_bus_power

SintPowerCase.get_load_buses_power

SintPowerCase.get_load_buses_power

SintPowerCase.get_load_buses_reactive_power

SintPowerCase.get_load_indices

SintPowerCase.get_n_buses

SintPowerCase.get_n_lines

SintPowerCase.get_n_os

SintPowerCase.get_power_injection_vector

SintPowerCase.get_primitive_admittance_matrix

SintPowerCase.get_susceptance_vector

SintPowerCase.is_load_bus

SintPowerCase.put_back_line!

SintPowerCase.set_bus_idx!

SintPowerCase.take_out_line!

SintPowerCase.to_ppc

SintPowerCase.contingency_matrix — Method

Creates a matrix that if subtracted to the admittance matrix implements a 
contingency.

Args:
    f_bus: from bus for the contingency
    t_bus: for the contingency.

SintPowerCase.create_random_states! — Method

Fills the loaddata and gendata matrices with random data.

It generates new operating states by addding a normally distributed
number to the load demand that is already in the case. The number is
given as a standard deviation multiplied with the current load at the bus.

SintPowerCase.get_admittance_matrix — Method

returns the admittance matrix of the system.

SintPowerCase.get_admittance_matrix — Method

Returns the incide matrix of the system as A'*Y_pr*A, where
A is the system incidence matrix and Y_pr is the primitive
admittance matrix.

SintPowerCase.get_all_buses_aggregated_power — Method

Aggregates the power for generators or loads for all buses

Inputs: df: Dataframe with either the power for all generators or loads. type: Active or reactive power.

SintPowerCase.get_bus_row — Method

Returns the row number of a bus given by id

SintPowerCase.get_dc_admittance_matrix — Method

get_dc_admittance_matrix(case)
Returns the admittance matrix for performing a dc power flow.
Inputs:
    case: The power system data.
    consider_status: If branch in service status should be considered.

SintPowerCase.get_gen_bus_power — Method

returns the total gen power at a bus.

SintPowerCase.get_gen_buses_power — Method

Returns the power for generator buses for os

SintPowerCase.get_gen_buses_power — Method

Returns the sum of production at generator buses.

SintPowerCase.get_gen_indices — Method

Return indices of the buses with generators.

SintPowerCase.get_id_idx — Method

Returns the row number of an element given by id

SintPowerCase.get_incidence_matrix — Method

get_incidence_matrix(case)::Array{Float64}
Returns the incidence matrix of a power system.

Inputs:
    case: The power system to get the susceptance vector for.
    consider_status: Whether or not branch in service status should be
    considered.

SintPowerCase.get_load_bus_power — Method

returns the total load power at a bus.

SintPowerCase.get_load_buses_power — Method

Returns the power demamd for load buses for os

SintPowerCase.get_load_buses_power — Method

Returns the sum of active load at load buses.

SintPowerCase.get_load_buses_reactive_power — Method

Returns the sum of reactive consumption at load buses.

SintPowerCase.get_load_indices — Method

Return indices of the buses with loads.

SintPowerCase.get_n_buses — Method

Returns the number of buses in the case.

SintPowerCase.get_n_lines — Method

Returns the number of lines in the case.

SintPowerCase.get_n_os — Method

Returns the number of oses in the case.

SintPowerCase.get_power_injection_vector — Method

Returns the power injection vector.

SintPowerCase.get_primitive_admittance_matrix — Method

Returns the primitive admittance matrix for a power system.
This is a matrix with dimensions BxB where B is the number of branches.
It is a diagonal matrix with the sum of the line and shunt impedances
on the diagonal.

SintPowerCase.get_susceptance_vector — Method

get_susceptance_vector(case)
Returns the susceptance vector for performing a dc power flow.

Inputs:
    case: The power system to get the susceptance vector for.
    consider_status: Whether or not branch in service status should be
        considered.

SintPowerCase.is_load_bus — Method

is_load_bus(case, id::String)
Returns true if the bus bus_id is a load.

SintPowerCase.put_back_line! — Method

Sets branch to out of service

SintPowerCase.set_bus_idx! — Method

Find the indices of the buses where the generators or loads are located

SintPowerCase.take_out_line! — Method

Sets branch to out of service

SintPowerCase.to_ppc — Method

Convert the case to the PYPOWER Case format.